VCI: return parser.py STDOUT/STDERR from Set()
Have frr-reload.py send logs to STDERR (via --stdout option) so that
any logs can be collected by the VCI component when it executes
parser.py. This means that frr-reload logs no longer end up in
/var/log/frr/frr-reload.log.

If parser.py (likely due to frr-reload.py) exits non-zero then any
collected logs are returned from the VCI Set(), replacing the generic
exited non-zero error message, and are displayed on the console.

Additionally, if the -debug argument is passed to the VCI component
the logs are always returned from Set(), even if parser.py exited
with a zero exit code.

In the usual case, we don't return logs if parser.py exits with a
zero exit code since frr-reload.py is not silent on success.

VRVDR-52211