Deviations for DANOS route-map configuration
Add the vyatta-policy-route-danos-v1 YANG module which contains
deviations to vyatta-policy-route-v1 to make it more suitable to
DANOS (and FRR as the routing stack).

The following route-map nodes are marked as "not-supported" and will
not be available to configure:
  - "match extcommunity exact-match"
  - "match ipv6 nexthop"
  - "match ipv6 peer"
  - "set delete-extcommunity"
  - "set prepend-as own-as"
  - "set extcommunity rt"
  - "set level"

Since "set extcommunity rt" is not-supported, remove its translation.
This translation is the same as "set add-extcommunity rt" which is
still supported.

Additionally a must statement is added to enforce mutual exclusion
between "match ip source-protocol" and "match ipv6 source-protocol".

The maximum insertion count for "set prepend-as last-as" is also set
to 10, matching the maximum value available in FRR.

VRVDR-52244