parser.py: Add SyntaxFile class
SyntaxFile represents a single syntax (translation) file on disk. A
load() method is implemented which returns the syntax dictionary as
defined by the JSON object in the file.

VyattaJSONParser.discover_syntax() generates a dictionary of SyntaxFile
objects, keyed by filename, and stored as the syntax_files instance
variable.

VyattaJSONParser.read_syntax_files() is updated to work on SyntaxFile
objects, after a call to discover_syntax().