parser.py: prioritise config before reading syntax
Prioritising the input configuration does not require the syntax
(translation) files to be loaded. Therefore perform prioritisation
immediately after loading the input config.

In future this change will allow the configuration parsing to be more
finely tuned by allowing only a subset of syntax files to be loaded
at any one time, effectively determining which parts of the input
configuration are translated.

We also make a change to the VyattaJSONParser constructor to ensure
that the syntax dictionary is always (empty) allocated. This is
required since it is referenced in the prioritize() call stack:

  depth_first_traverse()
    on_exit()
      retrieve_commands()
        syntax.get()