parser.py: Add steps support
This commit allows the parsing to be broken down into smaller logical
chunks through the new steps.json configuration file.

steps.json defines a JSON array of objects with each object
representing a parsing "step". To complete the parsing operation
steps are executed sequentially as defined in steps.json.

The steps.json configuration as defined in this commit results in the
same behaviour as before.

See the additions to scripts/frr/README.md in this commit for further
information.