steps.json: Apply static translations first
Previously in cases where static route configuration is applied in a
routing instance along with BGP (in default or non-default routing
instance) frr-reload.py could fail as the "vrf" line (of the
static route configuration) was interpreted as the vrf-policy command
of BGP mode.

The root cause is somewhat multi-faceted. The frr.conf file we
generate as input to frr-reload.py triggers a failure in the
`vtysh -m` processing (spawned by frr-reload.py), since due to the
above we attempt to execute "exit-vrf" when not in vrf mode, which
yields an unknown command error. See [1].

This can be addressed by updating the BGP translations to ensure that
each "router bgp" configuration block is terminated with an "exit"
line - `vtysh -m` will now happily mark the file.

However the issue still remains as frr-reload.py, following its
processing, does not retain these "exit" lines. As a result when it
goes to apply the configuration changes via `vtysh -f` it still
attempts to enter the "vrf" line of the static route configuration
while in BGP mode.

Since this breakage is caused by the ordering of the input config
to frr-reload.py we can workaround it by simply moving all static
route configuration to the start of the input config.

Having the static route configuration output prior to any protocol
configuration does also match the behaviour from vtysh "show run".

DAN-176

[1] https://github.com/FRRouting/frr/issues/7326