bgp.json: Improve peer-group translations
The neighbor should be assigned to a peer-group in bgp mode rather
than under each address family.

When the peer-group is assigned under an address family it is
represented in the FRR running configuration as though it were
configured in the parent bgp mode.

Therefore frr-reload.py determines a diff between the running and
desired configuration which results in un-necessary re-configuration,
and therefore peer flaps.

With the new translations we generate the peer-group assignment
configuration in bgp mode regardless of what address family it was
configured under (we first try ipv4-unicast, then ipv6-unicast).
This is OK since validation ensures that the same peer-group is
assigned for each address family of a given neighbor.

In this commit we also remove some unused (they expect a "neighbor"
node beneath each address family, which does not exist) translations
from the peer-group configuration trees.

DAN-212