debian/control: Add dependency on frr-snmp
This ensures FRR's SNMP support is included in DANOS images.
Manual configuration is currently required to enable SNMP support,
this change is simply a convenience to remove the need for users
to build/install the package themselves.

To enable SNMP support in FRR one must:
  * Edit /etc/frr/daemons and include `-M snmp` in the options for
    all desired & supported daemons (zebra, bgpd, ospfd, ospf6d).
  * Edit /etc/vyatta-routing/configs/steps.json and include the
    `agentx` config in the first step.

In future we could add a configuration data model to support
enabling/disabling SNMP support in FRR.

DAN-272