vci: defer FRR config until VRFs are provisioned
Currently we often attempt to reload the FRR configuration before
VRFs (routing instances) have been provisioned (due to commit
ordering - routing instance config is not handled by a dedicated
VCI component like our protocols config is).

This prevents the protocol configuration in these VRFs from being
correctly/fully applied in FRR, since it expects the VRFs to be
defined on the system before receiving associated configuration.

This issue can be observed:
 - when configuring a new routing instance, with protocols config,
   in a single commit
 - on boot

To address this we now, on receipt of new configuration to our VCI
component, defer the FRR configuration reload if the new configuration
references VRFs which are not yet provisioned.

To determine when these VRFs are provisioned on the system we
subscribe to the vyatta-routing-v1:instance-added notification.
The reload is triggered once we have been notified that all
referenced VRFs have been provisioned.

To allow us to ensure this ordering behaviour on system boot, a
further change is made to remove /etc/frr/frr.conf before FRR starts.
This prevents FRR from loading its own configuration when it starts,
which is possibly before VRFs are provisioned on the system.
Instead it will start with an empty configuration, and the desired
configuration will later be applied when the VCI component triggers
the reload.

DAN-317