Generate base FRR config on boot
c71e71fef5387719addd00d071a5b35a0e5c5db1 results in the required FPM
configuration not being applied at bootup until the VCI component
is triggered.

This is fine, except for the case where no configuration owned by
vyatta-frr-vci is present. In this case the VCI component is not
triggered and thus the FPM configuration is not applied.

When the FPM configuration is not applied zebra will not attempt to
connect to brokerd, and it thus sits there waiting for a connection.
The dataplane is therefore never programmed until/unless some
protocols configuration is applied.

To address this we add a oneshot service which runs before FRR, if
it does not have a configuration file, and runs the config parser
to generate the base config (via "empty" config input).

VRVDR-54511